Home

uCSV.jl Documentation

Functions

uCSV.read — Function.

read(fullpath;
     delim=',',
     quotes=null,
     escape=null,
     comment=null,
     encodings=Dict{String, Any}(),
     header=0,
     skiprows=Vector{Int}(),
     types=Dict{Int,DataType}(),
     isnullable=Dict{Int,Bool}(),
     coltypes=Vector,
     colparsers=Dict{Int,Function}(),
     typeparsers=Dict{DataType, Function}(),
     typedetectrows=1,
     skipmalformed=false,
     trimwhitespace=false)

Take an input file or IO source and user-defined parsing rules and return:

a Vector{Any} containing the parsed columns

a Vector{String} containing the header (column names)

Arguments

fullpath::Union{String,IO}

the path to a local file, or an open IO source from which to read data

delim::Union{Char,String}

whatever the dataset is being split on

default: delim=','

for CSV files

frequently used:

delim='\t'

delim=' '

delim='|'

may contain any valid UTF-8 character or string

quotes::Union{Char,Null}

the character used for quoting fields in the dataset

default: quotes=null

by default, the parser does not check for quotes.

frequently used:

quotes='"'

escape::Union{Char,Null}

the character used for escaping other special & reserved parsing characters

default: escape=null

by default, the parser does not check for escapes.

frequently used:

escape='"'

double-quotes within quotes, e.g. "firstname ""nickname"" lastname"

escape='\\'

note that the first backslash is just to escape the second backslash

e.g. "firstname \"nickname\" lastname"

comment::Union{Char,String,Null}

the character or string used for comment lines in your dataset

note that skipped comment lines do not contribute to the line count for the header (if the user requests parsing a header on a specific row) or for skiprows

default: comment=null

by default, the parser does not check for comments

frequently used:

comment='#'

comment='!'

comment="#!"

encodings::Dict{String,Any}

A dictionary mapping parsed strings to desired julia values

if your dataset has booleans that are not represented as "true" and "false" or missing values that you'd like to read as nulls, you'll need to use this!

default: encodings=Dict{String, Any}()

by default, the parser does not check for any reserved fields

frequently used:

encodings=Dict{String, Any}("" => null)

encodings=Dict{String, Any}("NA" => null)

encodings=Dict{String, Any}("N/A" => null)

encodings=Dict{String, Any}("NULL" => null)

encodings=Dict{String, Any}("TRUE" => true, "FALSE" => false)

encodings=Dict{String, Any}("True" => true, "False" => false)

encodings=Dict{String, Any}("T" => true, "F" => false)

encodings=Dict{String, Any}("yes" => true, "no" => false)

... your encodings here ...

can include any number of String => value mappings

note that if the user requests quotes, escapes, or trimwhitespace, these requests will be applied (removed) the raw string BEFORE checking whether the field matches any strings in in the encodings argument.

header::Union{Integer,Vector{String}}

The line in the dataset on which to parse the header

note that commented lines and blank lines do not contribute to this value e.g. if the first 3 lines of your dataset are comments, you'll still need to set header=1 to interpret the first line of parsed data as the header.

default: header=0

no header is checked for by default

frequently used:

header=1

skiprows::AbstractVector{Int}

A Vector or Range (e.g. 1:10) or rows to skip

note that this is 1-based in reference to the first row AFTER the header. If header=0 or is provided by the user, this will be the first non-empty line in the dataset. Otherwise skiprows=1:1 will skip the header+1-nth line in the file.

default: skiprows=Vector{Int}()

no rows are skipped

frequently used:

skiprows=6:typemax(Int)

read only the first 5 lines of the dataset

types::Union{Type, Dict{Int, Type}, Dict{String, Type}, Vector{Type}}

declare the types of the columns

scalar, e.g. types=Bool

scalars will be broadcast to apply to every column of the dataset

vector, e.g. types=[Bool, Int, Float64, String, Symbol, Date, DateTime]

the vector length must match the number of parsed columns

dictionary, e.g. types=("column1" => Bool) or types=(1 => Union{Int, Null})

users can refer to the columns by name (only if a header is provided or parsed!) or by index

default:

types=Dict{Int,DataType}()

column-types will be interpreted from the dataset

built-in support for parsing the following:

Int

Float64

String

Symbol

Date – only the default date format will work

DateTime – only the default datetime format will work

for other types or unsupported formats, see colparsers and typeparsers

isnullable::Union{Bool, Dict{Int, Bool}, Dict{String, Bool}, Vector{Bool}}

declare whether columns should have element-type Union{T, Null} where T

scalar, e.g. isnullable=true

scalars will be broadcast to apply to every column of the dataset

vector, e.g. isnullable=[true, false, true, true]

the vector length must match the number of parsed columns

dictionary, e.g. isnullable=("column1" => true) or isnullable=(17 => true)

users can refer to the columns by name (only if a header is provided or parsed!) or by index

default: isnullable=Dict{Int,Bool}()

column-types are only nullable if null values are detected in rows 1:typedetectrows.

coltypes::Union{AbstractVector, Dict{Int, AbstractVector}, Dict{String, AbstractVector}, Vector{AbstractVector}}

declare the type of vector that should be used for columns

should work for any AbstractVector that allows push!ing values

scalar, e.g. coltypes=CategoricalVector

scalars will be broadcast to apply to every column of the dataset

vector, e.g. coltypes=[CategoricalVector, Vector, CategoricalVector]

the vector length must match the number of parsed columns

dictionary, e.g. coltypes=("column1" => CategoricalVector) or coltypes=(17 => CategoricalVector)

users can refer to the columns by name (only if a header is provided or parsed!) or by index

default: coltypes=Vector

all columns are returned as standard julia Vectors

colparsers::Union{Function, Dict{Int, Function}, Dict{String, Function}, Vector{Function}}

provide custom functions for converting parsed strings to values by column

scalar, e.g. colparsers=(x -> parse(Float64, replace(x, ',', '.')))

scalars will be broadcast to apply to every column of the dataset

vector, e.g. colparsers=[x -> mydateparser(x), x -> mytimeparser(x)]

the vector length must match the number of parsed columns

dictionary, e.g. colparsers=("column1" => x -> mydateparser(x))

users can refer to the columns by name (only if a header is provided or parsed!) or by index

default: colparsers=Dict{Int,Function}()

column parsers are determined based on user-specified types and those detected from the data

typeparsers::Dict{Type, Function}

provide custom functions for converting parsed strings to values by column type

note user must specify column types for this to have the intended effect, as the parser uses the default type-parsers for detecting column type.

default: colparsers=Dict{DataType, Function}()

column parsers are determined based on user-specified types and those detected from the data

frequently used:

typeparsers=Dict(Int => x -> parse(Float64, replace(x, ',', '.'))) # euro-style floats!

in combination with types to specify which columns to apply the parsers to.

typedetectrows::Int=1

specify how many rows of data to read before interpretting the values that each column should take on

commented, skipped, and empty lines are not counted when determining which rows are used for type detection, e.g. setting typedetectrows=10 and skiprows=1:5 means type detection will occur on rows 6:15

skipmalformed::Bool=false

specify whether the parser should skip a line or fail with an error if a line is parsed but does not contain the expected number of rows

default: skipmalformed=false

malformed lines result in an error

trimwhitespace::Bool=false

specify whether should extra whitespace be removed from the beginning and ends of fields.

leading and trailing whitespace OUTSIDE of quoted fields is trimmed by default.

trimwhitespace=true will also trim leading and trailing whitespace WITHIN quotes

source

uCSV.write — Function.

function write(fullpath;
               header=null,
               data=null,
               delim=',',
               quotes=null,
               quotetypes=AbstractString)

write a dataset to disk or IO

Arguments

fullpath::Union{String, IO}

the path on disk or IO where you want to write to.

header::Union{Vector{String}, Null}

the column names for the data

default: header=null

no header is written

data::Union{Vector{<:Any}, Null}

the dataset to write to disk or IO

default: data=null

no data is written

delim::Union{Char, String}

the delimiter to seperate fields by

default: delim=','

for CSV files

frequently used:

delim='\t'

delim=' '

delim='|'

quotes::Union{Char, Null}

the quoting character to use when writing fields

default: quotes=null

fields are not quoted by default, and fields are written using julia's default string-printing mechanisms

quotetypes::Type

when quoting fields, quote only columns where coltype <: quotetypes

default: quotetypes=AbsractString

only the header and fields where coltype <: AbsractString will be quoted

note that columns of Union{<:coltype, Null} will also be quoted, for cases where columns that you desire to be quoted also have missing values.

frequently used:

quotetypes=Any

quote every field in the dataset

source

function write(fullpath,
               df;
               delim=',',
               quotes=null,
               quotetypes=AbstractString)

write a DataFrame to disk or IO

Arguments

fullpath::Union{String, IO}

the path on disk or IO where you want to write to.

df::DataFrame

the DataFrame to write to disk or IO

delim::Union{Char, String}

the delimiter to seperate fields by

default: delim=','

for CSV files

frequently used:

delim='\t'

delim=' '

delim='|'

quotes::Union{Char, Null}

the quoting character to use when writing fields

default: quotes=null

fields are not quoted by default, and fields are written using julia's default string-printing mechanisms

quotetypes::Type

when quoting fields, quote only columns where coltype <: quotetypes

default: quotetypes=AbsractString

only the header and fields where coltype <: AbsractString will be quoted

note that columns of Union{<:coltype, Null} will also be quoted, for cases where columns that you desire to be quoted also have missing values.

frequently used:

quotetypes=Any

quote every field in the dataset

source

Manual

Getting Started

Headers

Reading into DataFrames

Delimiters

Missing Data

Declaring Column Element Types

Declaring Column Vector Types

International Representations for Numbers

Custom Parsers

Quotes and Escapes

Skipping Comments and Rows

Malformed Data

Reading data from URLs

Reading Compressed Datasets

Common formatting issues

Benchmarks